adjusted in correct folders (auth) and added authgate in main
