+++ lib/main.dart
@@ -1,18 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:lumi_learn_app/screens/start/main_start.dart';
 import 'package:google_fonts/google_fonts.dart';
+
 import 'controllers/navigation_controller.dart';
-import 'screens/home/home_screen.dart';
-import 'screens/social/social_screen.dart';
-import 'screens/profile/profile_screen.dart';
-import 'widgets/bottom_nav_bar.dart';
-import 'utils/constants.dart';
-import 'screens/launch/launch_screen.dart';
-import 'screens/start/login/login_screen.dart';
-import 'screens/start/sign/signup_screen.dart';
+import 'screens/auth/auth_gate.dart'; // Your AuthGate widget
 
 void main() {
+  // Initialize your controllers and dependencies
   Get.put(NavigationController());
   runApp(const MyApp());
 }
@@ -26,38 +20,14 @@ class MyApp extends StatelessWidget {
       title: 'Lumi Learn',
       theme: ThemeData(
         primarySwatch: Colors.blue,
-        // make dark theme the default
         brightness: Brightness.dark,
         textTheme: GoogleFonts.poppinsTextTheme(),
       ),
-      home: const MainScreen(),
-    );
-  }
-}
-
-class MainScreen extends StatelessWidget {
-  const MainScreen({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final NavigationController navigationController = Get.find();
-
-    return Scaffold(
-      body: Obx(
-        () => IndexedStack(
-          index: navigationController.currentIndex.value,
-          children: [
-            LaunchScreen(),
-            MainStartScreen(),
-            LoginScreen(),
-            SignupScreen(),
-            const HomeScreen(),
-            const SearchScreen(),
-            const ProfileScreen(),
-          ],
-        ),
-      ),
-      // bottomNavigationBar: const BottomNavbar(),
+      // Use AuthGate as the home widget to handle flow control:
+      home: const AuthGate(),
+      // Optionally, if you prefer named routes:
+      // initialRoute: Routes.authGate,
+      // getPages: getPages,
     );
   }
 }
